modified 0-print_listint file

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -6,7 +6,7 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	unsigned int count;
+	size_t count;
 
 	if (h == NULL)
 		return (0);
